requirements.txt
spotipy==2.4.4
google-auth==1.7.1
google-cloud==0.34.0
googleapis-common-protos==1.6.0
google-cloud-texttospeech==0.5.0
pyserial==0.3.5